refactor(analysis): extract boundary placement into its own module
segment_generation.py mixed two jobs: generating candidate segments from cut
points, and deciding where a cut may legally land given protected audio. The
second half moves to boundary_placement.py, which leaves segment_generation
at 554 lines and gives the shared helpers a home where they are not private
to a module that only half-owns them.

_gaps_between and _cap_end_to_gap lose their underscore -- unified_analyzer
already imported the former across module lines.
